+++ features/auth/login/index.ts
@@ -1,0 +1,1 @@
+export { LoginForm } from './login.ui';
