+++ features/auth/register/index.ts
@@ -1,0 +1,1 @@
+export { RegisterForm } from './register.ui';
